Приведение пути (мапинг) во всех рест-контроллерах в crm к единому виду Объединение двух классов с похожим функционалом  и маппингом в один
