+++ src/main/java/com/ewp/crm/controllers/rest/ClientHistoryRestController.java
@@ -1,5 +1,6 @@
 package com.ewp.crm.controllers.rest;
 
+import com.ewp.crm.models.ClientHistory;
 import com.ewp.crm.models.dto.ClientHistoryDto;
 import com.ewp.crm.service.interfaces.ClientHistoryService;
 import org.slf4j.Logger;
@@ -9,10 +10,11 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
-@RequestMapping("/client/history/rest")
+@RequestMapping("/rest/client/history")
 public class ClientHistoryRestController {
 
 	private static final Logger logger = LoggerFactory.getLogger(ClientHistoryRestController.class);
@@ -31,6 +33,24 @@ public class ClientHistoryRestController {
 	public ResponseEntity getClientHistory(@PathVariable("clientId") long id, @RequestParam("page")int page, @RequestParam("isAsc")boolean isAsc) {
 		List<ClientHistoryDto> clientHistory = clientHistoryService.getAllDtoByClientId(id, page, pageSize, isAsc);
 		return ResponseEntity.ok(clientHistory);
+	}
+
+	@GetMapping("/getEmailHistory/{clientId}")
+	public ResponseEntity getClientEmailMessages(@PathVariable("clientId") long id) {
+		List<ClientHistory> clientHistories = clientHistoryService.getByClientId(id);
+		clientHistories.removeIf(x -> !x.getType().name().equals("SEND_MESSAGE"));
+		List<String[]> response = new ArrayList<>();
+		for (ClientHistory clientHistory: clientHistories) {
+			String time = clientHistory.getDate().toString();
+			String author = clientHistory.getMessage().getAuthorName();
+			if (author == null || author.isEmpty()) {
+				author = "неопределен";
+			}
+			String content = clientHistory.getMessage().getContent();
+			String[] messageData = {time, author, content};
+			response.add(messageData);
+		}
+		return ResponseEntity.ok(response);
 	}
 
 }
